Initial commit
Basically just getting the server running for the first time. It's able
to start up a headless browser instance and login to the form
sucessfully. Now just requires actually scraping html and css, would be
nice to actually take the colours assigned to the periods so I think
I'll try and add that as a feature.